feat(conf) add 'lua_ssl_verify_depth' property
When specifying a certificate for datastore connections, it can be
required to adjust the verification depth in the certificates chain.

This adds the `lua_ssl_verify_depth` directive when *and only when* a
certificate is configured for Cassandra connections.

Once PostgreSQL SSL connections will be supported, we will handle better
the `lua_ssl_trusted_certificate` directive as well as the verify depth.

* New property: new default + added and documented it in
  `kong.conf.default`
* new test

Implements #1395