refactor(cli) cleaner from-to dependencies
I believe only 1 argument is necessary to version.lua when no range is
needed (as it seemds to be indicated in its documentation as well as
previous usage of this module in the CLI), but if we wish to maintain
explicit version ranges, I believe this is cleaner.

Additionally, show a better error message for incompatible 'resty'
versions.