fix(router) don't append trailing slashes unconditionally
We now let the consumer of an API decide whether or not a trailing slash
should be required. If the requests' URI ends up being "/", (post URI
stripping), we then defer the decision to the `upstream_url` configured
on the API. The details of the implementation are discussed in #2211.

Fix #2211