fix(admin) CORS ACAC header cannot receive false
Remove the ACAC header set to `false` in the Admin API.

This raises an interesting concern about the Admin API: we do not allow
requests with credentials to be made to it (since ACAC is not sent with
`true`). This might warrant an update and some testing in the future.