test(*) include plugin fixture Lua path
The custom plugins used for testing were invoked in different ways.
This includes the path to the plugins in both the test engine (busted)
and in the test Kong instance. This way the plugins can simply
be inserted through the faster, and more reliable, dao-helpers instead
of having to do it through an http request.