refactor(core) use new invalidations framework
We take advantage of the new `kong.cache` module to build a better
configuration entities caching mechanism, with invalidation events that
do not use the locally running Serf agent.

This refactors all caching branches of the core entities (APIs,
Consumers, Plugins, Credentials...) and provides more reliable and
performant integration tests.

We purposefully do not remove any Serf-related code. At this stage, 2
invalidation mechanism are running side-by-side.

Additionally, this takes advantage of `lua-resty-worker-events`
supporting inter-worker and intra-worker communication to build the
foundation of a "event-driven" handlers (such as being able to
subscribe to CRUD operations in the configuration entities, for future
plugins and such). `lua-resty-worker-events` is - unfortunately -
*eventually consistent* across workers, and will *eventually* either be
improved, or replaced with a consistent - and more performant -
implementation (one is currently being developed as we speak, but future
OpenResty work may also open the way to proper IPC via cosockets).