fix(cors) send '*' in ACAO when 'conf.origins' is empty
This fixes a regression introduced in 0.10.1, where the migration would
set the new `conf.origins` field to an empty table, but the code would
only set ACAO to `*` if the table was `nil`, and not empty.

We are reluctant to update the migration due to the immutable nature
that they should carry (in order to avoid divergence between users who
already ran them, and users who will run it after this patch), so this
changes the business logic to be more robust instead, which we probably
want anyways.

This regression was documented post-release, in #2517.