refactor(plugins) use new invalidations framework
Like our previous work to use the new `kong.cache` module for core
entities instead of `database_cache.lua`, this takes care of it for all
entities used in plugins (theoretically not part of Kong core). It is
also considered a separate unit of work for the sake of limiting the
patch size of a single commit, and ease future archaeology work.