hotfix(ngxconf) ensure init/init_worker logs show up
This is a follow up fix to
23585aad6db6bc4dfc7436965369ed2199e3d44a.

Since the `anonymous_reports` configuration property enables
an `error_log` directive with an `error` level, most of the
logs from the `init_worker` phase would be ignored.

This adds an `error_log` directive in the `http` block to
ensure we log `init_worker` logs at the proper level.

We also set the proper path (from the new configuration value)
to the master process's error logs. We use the proxy value as
we consider Kong's proxy server to be the main server on a Kong
instance.

From #2564