feat(admin) add Admin API post-processing utility
CRUD results can now be post-processed by `crud_helpers.lua` to apply
any necessary modification to the generated entities (or list of
entities). Ex: remove a value, JSON-decode a JSON-encoded value, etc...

We take advantage of this post-processing to send the created API
entities to the anonymous reports, which means we need a deep copy of
said entity, as it is modified by the reports module.