feat(router) return matched host/uri/method for plugins consumption
We now return any value that matched back to the core so that plugins
can make use of it.

This also improves the router performance by moving the URI
stripping/trailing slash/preserve host header to be properly cached by
the router.

We also simplify the return values of the router by providing a table
instead of yet one more return value.