hotfix(healthchecks) disable circuit-breaker functionality by default
To avoid unexpected behaviors for users, this makes all health counter
thresholds to be zero by default, effectively making the circuit-breaker
functionality of passive healthchecks to be disabled by default.
Users need to opt-in by supplying threshold values to enable the
cut-off of targets based on analysis of ongoing traffic.

This commit requires functionality introduced in lua-resty-healthcheck 0.3.0.