hotfix(runloop) ensure delayed response short-circuits plugin execution
Prior to c6bde1a, the `responses.send` method was effectively an "abort"
function, causing a plugin phase to stop executing. With the delayed
execution of the actual response sending, this feature of
`responses.send` was removed, breaking expectations of plugins which
assumed that a function would stop running if a response was sent with
`responses.send`. This PR restores the original behavior of aborting the
ongoing phase of the plugin, while still retaining the benefits of
c6bde1a, meaning that further phases and plugins still run.

This issue was originally spotted by @nateslo for the key-auth plugin,
and the precise issue was mapped out by @thibaultcha and originally
described in PR #3413.

From #3146

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>