fix(runloop) propagate Lua errors from coroutine.wrap()
It seems like ngx_lua's `corouting.wrap()` does not propagate Lua errors
to the parent thread like PUC-Lua or LuaJIT's coroutine do - this seems
to be an oversight for which a fix is currently being designed.

This ensures Lua errors happening inside plugins' access phase propagate
to the ngx_lua entry thread.