fix(balancer) async initialization of balancer objects
When upstreams include targets using hostnames instead of IPs, the
initialization of balancer objects needs to perform DNS resolution, which
involves opening a socket, which is not available in the init_worker phase.

This patch performs balancer initialization in a timer, which makes
initialization asynchronous to worker initialization. For this reason,
it adds a locking system to create_balancer to avoid race conditions
when requests happen while the balancers are being initialized.

The locking system uses a simplified version of `resty.lock` which does
not use the shm, since balancers are per-worker and we are concerned only
with intra-worker concurrency control.