fix(balancer) unregister healthchecker callback immediately
When a healthchecker is stopped, unregister its callback immediately,
instead of waiting for the garbage collector to kick in and
automatically reclaim it (as per the behavior of `register_weak`).

This does not modify the logic of health checks, but prevents
bogus error messages from appearing in the error logs.

Thanks to @aledbf for the report!