fix(cache) do not use callback as hard-error wrapper keys
Part of the fixes for #3277 for a regression introduced in
de6390e34c5301aa4c965d52f08e00ab81f91efa.

While the soft_to_hard error wrapper tries its best to be GC-friendly by
not keeping references on the DB cache callbacks as they are used as keys
to index their corresponding soft_to_hard_error wrappers.

Unfortunately, because the created closure holds a reference to the
callback itself, said callback cannot be garbage-collected.

This is especially troublesome for callbacks that are generated on the
fly, like so:

    local cache, err = cache:get("key", opts, function()
        return "cached value"
    end)

The above anonymous function should be garbage-collected, but the
soft_to_hard_error still holds a reference to it.
This example could be observed prior to the parent commit, in which we
replaced the cache callback to not create a function on the fly anymore.

The drawback of this change is a larger memory footprint, the same
cache callback will result in several copies of the same
soft_to_hard_error function, if they are invoked with several caching
keys.