fix(*) ensure rate-limiting plugins select proper Redis database
Since every instance of both rate-limiting plugin can share Redis
connection pools, and because each plugin can be configured with a
different Redis database, we must ensure that as each plugin gets
processes in every request, it uses the Redis database it was configured
for. We thus force database selection (see comment in code).

Fix #3292
From #3293

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>