fix(cli) redirect nginx stdout to kong CLI when nginx_daemon = off
The previous behavior (which was the same for both nginx_daemon =
on|off) was to call 'pl_utils.executeex()', which redirects
stdout/stderr to tmp files.

When nginx is configured with logs doing to /dev/stdout+/dev/stderr, and
nginx_daemon = off, then those files will receive all output. Which
brings two issues:

1. the output is hidden from the 'kong start' stdout/stderr, which is
not the expected behavior
2. the tmp files will slowly grow and fill up the disk space - they also
cannot be rotated

From #3297

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>