feat(jwt) support for key_claim_name in header
The JWT plugin previously only supported setting the
config.key_claim_name parameter to a claim in the body and not the
header.

Many identity providers who use JWKS (such as Amazon Cognito) support
signing their tokens with more than one public key. This results in the
need to support the "kid" (key ID) claim which identifies the public key
required to validate the signature of the token. Due to the JWKS
defining multiple public keys, the standard "iss" (issuer) claim is not
unique. The "kid" claim is typically set in the header of the token and
therefore requires also checking the token header for claims.

The change that was made is fully backwards compatible due to keeping
the functionality that checks the token body claims first and if the
config.key_claim_name is not found there it then checks the header. An
additional test case was also written to confirm that setting
config.key_claim_name to a claim that is in the header succeeds.

JWKS standard:
  https://tools.ietf.org/html/rfc7517#section-5

Fix #2858
From #3313

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>