fix(dao) prevent self_check from receiving ngx.null
Because `self_check` for plugins currently does a new DB `:select()` to
check for foreign entities, if the `service_id` or `route_id` fields are
`ngx.null` (from a PATCH with JSON NULL), then Cassandra produces an
error:

[Invalid] Invalid null value in condition for column id

And PostgreSQL as well, for another reason:

runtime error: ./kong/dao/schemas/plugins.lua:101: attempt to concatenate field 'service_id' (a userdata value)

We should avoiding giving `ngx.null` to `self_check` since the entire
old schema was never designed to play nice with it, and because external
plugins might be broken if we do so. This change improves
backwards-compatibility.

Fix #3308
From #3355

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>