tests(helpers) ensure wait_until runs with an updated time
Update the internal `ngx` clock before calculating the start time of the
`wait_until` loop. This also allows moving the `ngx.sleep` of the loop after
the `pcall` (without the `ngx.update_time` at the top, some tests would break
if `ngx.sleep` was done after the `pcall` as they were implicitly assuming
that they ran with an updated clock).

Also shorten the wait in the poll loop while we're at it!