feat(schema) friendlier behavior for at_least_one_of
Make the behavior of the validation of routes friendlier:

Without this patch, if a user performs a PATCH request
without giving explicit values to all three of `methods`,
`hosts` or `paths`, we reject the request because we
need to cross-check all three fields to ensure that
at least one of those is not-null. If only one of those
is to be set, the user needs to explicitly set
the other two to `null` in the request.

With this patch, if a user performs a PATCH request giving
one non-empty value for at least one of `methods`, `hosts`
or `paths`, we now accept it even if the other fields are
missing, because this request cannot put these fields in
an invalid state.

This was implementing in a very simple way, by adding a
new internal property to `entity_checkers`, called
`run_with_missing_fields`. Currently, the only entity
checker that can run with missing fields is `at_least_one_of`,
which is used by the Route entity.

---

feat(schema) de-hardcode entity checker error messages

Makes it easier to customize them, and in the future internationalize
them.

---

feat(schema) update-aware messages for entity checks

Monkey-patch some error messages to make it clearer why they apply
during an update. This is a bit ugly, but it avoids propagating
update-awareness all the way down to the entity checkers (which would
otherwise defeat the whole purpose of the abstracted entity checker
mechanism, created because we don't want low-level logic to be dependent
on `is_update` flags, which in the past led to relevant validations
being overly skipped during updates).

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>