feat(config) add 'headers' property to replace '*_tokens'
As new headers are introduced by Kong's core, instead of introducing a
config option for each and every header or set of headers, an array of
these values can be now specified using the `headers` config option.

This offers a fine grained controlled and does not blow up the number of
configuration options.

The goal here is to move towards a  simpler and easier to understand
configuration, similar to 1b9976f60 (#3147).

From #3300

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>