fix(balancer) avoid needless rebuild of balancer when adding target
This bug was introduced when the loop to find `last_equal_index`
was changed from looping the new targets list into looping the
old targets list. Looping the old targets list is more efficient,
but the logic for checking the index needed to be updated.

Includes a regression test that verifies that adding a target
works and does not rebuild the balancer.