feat(conf) inject 'prometheus_metrics' shm if needed
Prometheus plugin needs a dedicated shm to store metrics of Kong and
it's upstream. This patch sets the injects the requried shm into NGINX
template if the plugin is available for use.  The size chosen for the
shm is 5 MB, which is more than sufficient to store metrics related to a
few hundred services (tested for upto 200 services).

This is a temporary workaround until there is an API exposed from the
core which allows specifying custom shared dicts for plugin.  When such
an API is introduced, the changes introduced by commit should be
reverted.

From #3547