fix(schema) fix validation of nested records in metaschema
Includes a regression test in the metaschema testsuite, and a testcase in the
schema testsuite (which demonstrates that the nested records issue was a
metaschema-only problem).

Includes some minor corrections to the metaschema specification (detected
using the metaschema to attempt to validate itself... unfortunately the schema
validator doesn't implement a type system strong enough to fully validate the
metaschema without falling into an infinite loop -- in fact, the metaschema
itself "cheats" and performs some of its validation of schemas using the
`check` escape-hatch function to recurse into records, as we see in this
commit).

Still, the metaschema is useful enough to catch syntax errors in schema
definitions, so it's good to keep it around.