chore(makefile) auto-convert Windows line-endings (#3594)
Git clients on Windows tend to convert line endings to Windows
format, such that Windows editors can be used (is reverted on commit).
This doesn't work if mounting those files from a Unix machine for
development for example (Kong-Vagrant). Lua is agnostic to the format,
but the Unix shell is not. Hence only the scripts in ./kong/bin are
affected.

This change adds a Makefile target to easily fix up the files back to
Unix format. This helps Windows users working on the Kong code
and custom plugins.