feat(db) generic cache key support in new DB
This combines the fast-path behavior of the old DAO's `cache_key`
and the generic behavior of the old DAO's `entity_cache_key`.
Having it in the new DAO allows us to handle cache keys in the
`crud` events propagation system in a general manner, while the
fast path avoids us from degrading performance for the plugin
iterator.