feat(db) implement general cascade-delete for new DB
This adds general cascade delete support in the new DB.
Postgres does it natively. This commits adds the Cassandra
implementation, reusing the iteration logic via
`kong.db.iteration`.

This removes the custom cascade-delete code that was
present in the Consumers custom DAO, and in the custom
Cassandra strategies for Routes and Services.

This allows us to remove `kong.db.strategies.cassandra.routes` entirely. When
all custom DAO entities from the various plugins are ported to the new DB, it
will be possible to remove `kong.db.strategies.cassandra.services` as well.