feat(schema) make records non-nullable by default
Make records in Entities non-nullable by default, so that they return their
full structure on API queries. The idea of returning explicit nulls in API
queries was to allow users to see what are the available fields. In nested
structures, for that to happen we need to make records non-nullable, even if
their individual fields are.

A field of the `record` type in an Entity can still be made nullable if
desired, with an explicit `nullable = true` property.

This is done at the Entity level because the MetaSchema uses nullable
records extensively.