perf(plugins-iterator) only execute iterator on configured plugins
Reduce the number of invocations of the plugins iterator by attempting
to load configurations only for plugins we know are configured on the
cluster.

A map of distinct names of configs plugins is maintained by iterating
over each row in the plugins table. Plugins CRUD events are handled and
delete a node cache entry representing the current version of the
plugins map; in the request path this value is used to determine if a
rebuild of the map is necessary. This logic mimics that of the router
rebuild mechanism. Additionally, plugin maps rebuilds are wrapped in a
worker-level mutex (implemented via ngx.semaphore); this prevents
stampeding database connections during rebuild events in
high-concurrency situations.

From #3794

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>