feat(schema) do not run custom_entity_check if fields are invalid
For the generic checkers we had so far, running them on invalid fields and
reporting everything at once was okay, and avoided the "get one error, fix
thing, get another error" experience. But for a custom checker, it's safer to
only run it on valid fields (especially thinking of making things less
error-prone for third-party plugin authors).

The commit is lengthy because the `errors` table is now passed along
and is used to check whether there are recorded errors by the time
entity checkers run; a side effect of this is that the code is now
simpler, and produces fewer intermediate tables.