tests(*) fix some failing test cases due to lua-cassandra bump
Starting with lua-cassandra 1.3.3, we log a warning message when the
`rpc_address` of nodes is "bind all" (as with other Datastax drivers).

Updating the Travis-CI ccm scripts so that it binds on the local
interface has been considered, but since nothing guarantees that other
existing tests setups out there don't make the same mistake, this
approach is safer.