fix(schema) unify `required` and `nullable` attributes
The public notation for marking that a field in a Schema cannot be ngx.null is
`required`. However, we also had a second notation, `nullable`, which was used
only internally.

We used the `nullable` notation for implementing auto-setting of default
values when a value arrived as `NULL` from the database but was `nullable =
false` and had a `default` attribute.

These `nullable` annotations were being auto-inserted in a few places,
most importantly in the auto-converter of old-format schemas, so that
record values expanded properly.

Since `nullable` was used only internally in generated situations, the
MetaSchema had no knowledge about it (thereby hiding from the user
the existence of both `required` and `nullable`, which was definitely
a code smell).

This commit removes `nullable` and makes the `required` attribute take over
its duties, meaning that now new public fields can declare `default`
and `required` attributes and avoid the need for migrations, as was
originally intended (making this commit a "fix" rather than a mere
internal "refactor").

In practice, this is needed for the new `run_on` attribute.
The alternative would be to expose "nullable" to the MetaSchema,
a path that, if followed, would haunt us for a long time.

This commit adds a few extra tests.