fix(plugins) don't call redis:select when not necessary (#3973)
In rate-limiting and response-ratelimiting, a redis:select is called when connection is reused, even though the current configured redis database is 0, to clear the previously selected database. This will be problem when for some use cases like twemproxy where select is not supported, and sending `select` call will cause the connection to be closed.

The patch uses a host+port+database based redis connection pool, so that we always know the current connection is using correct database.