chore(conf) update default C* policy to 'RequestRoundRobin'
Introduced in 0.14, this new LB policy has been used in the wild and
should be considered stable. Since, in the context of Kong, this LB
policy has no downside and can potentially significantly reduce latency,
it is appropriate to use it as the default policy for 1.0. This ensures
that Kong is optimally configured from the get-go, as several complaints
about C* queries latency could have been mitigated by this policy
(although it is not the largest factor in those complaints).

By ensuring a given request reuses an existing connection, we avoid
additional handshake time (greatly worsened by SSL and authentication if
configured), and ensure consistency of data throughout the request.

From #4004