tests(unit) meaningful args in ngx.semaphore mock
While modifying the runloop router rebuild for a subsequent commit by
increasing/reducing the number of resources granted back to the
semaphore, the tests kept passing.

We now properly test that the semaphore resources management is tested
according to the runtime code's arguments. In ngx.semaphore, the default
argument for `post()` is `1`.