feat(pdk) pretty-print long and/or multiline logs with kong.log.inspect
The `kong.log.inspect` function is designed as a development tool for
inspecting Lua objects, many of which are complex tables. It leverages
@kikito's `inspect` library, which pretty-prints tables in a nice
multi-line indented presentation.

This commits aims to preserve that presentation, by splitting multiline
logs done with `kong.log.inspect` into multiple nginx log invocations
and pretty-printing them with spaces and a border. This results in
much easier to read tables, and allows you to read the entire contents,
since the logs have a short limit per log call.

The `kong.log.inspect` function is designed as a development-time tool,
so there are no production-time concerns about performance or printing
over-long messages.