fix(core) ensure a plugin created with PUT gets enabled
When a plugin is newly created with PUT (e.g. with
`/plugins/{uuid}`), this needs to trigger a regeneration
of the plugins map.

Our `dao:upsert` operation triggers an `"update"` CRUD event.
This change is the conservative approach to fixing the problem,
instead of changing the CRUD event and ensuring that all
consumers of CRUD events handle a new `"upsert"` event.

Includes a regression test.

Fixes #4191.