fix(conf) use same opts when reading default conf and kong.conf
This change allows for storing JSON strings in the defaults template.
Without the `smart = false`, pl_config is SO smart, that is converts to
the stringified JSON to a table and then `tostring` it in .kong_env as
a table [some hex address].

The addition of `list_delim = "_blank_"` is simply to match the entire
options object passed to the `pl_config.read` call of the user-provided
conf file.

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>