hotfix(db) allow plugins on empty services (#4312)
Before this change, assigning a plugin to an empty service (with no routes) would result in an “no compatible routes found in service” error. That is considered a breaking change (for things like automatic setups) and has been altered now: when a plugin is assigned to a service, and the service has no routes, then there will be no error. The error will still be raised when the service has at least one route, but none of its routes are compatible.