fix(db) fix target cleanup
Target cleanup was happening in the wrong order. This was not a problem for
the target being inserted (because cleanup happened before insertion), but it
would cause the other targets of the list to be corrupted in case they had a
target history themselves.

Since cleanup happened before insertion (meaning before a weight=0 entry was
inserted upon deletion), this meant that deleting targets never truly emptied
the table -- this meant that Cassandra could not delete upstreams, because the
cascade-delete annotation was missing from the entity schema. Merely adding
`on_delete = "cascade"` to `targets.upstream` would fix the problem in #4317
illustrated by the regression test added to
spec/02-integration/04-admin_api/07-upstreams_routes_spec.lua but would not
fix the other problem, illustrated by the test added to
spec/02-integration/04-admin_api/08-targets_routes_spec.lua.

Fixes #4317.