fix(*) quick fix lua_pack as it modifies standard lua string library
Lua 5.3 comes with `string.pack` and `string.unpack` so there are
libraries that try to see if they are in place, like @daurnimator's
`lua-http` library. The `lua_pack` modifies standard `string`
library, and adds `pack` and `unpack` functions to it which are
incompatible with Lua 5.3 versions of them. `lua-http` on the other
hand thinks it is getting the 5.3 ones, and then we see things
crashing.

Another PR with removal of `lua_pack` dependency will follow.

This is based on `next` so that I can rebase #4286 on top of `next`
when this is merged.