fix(schema) fix conundrum in self-validation of `default` attribute
In the MetaSchema, the `default` attribute has a special type, "self",
which is the type of the field in which it is itself present.
(This special type is not available for plugin/entity authors.)

The MetaSchema exists so that we can vet schemas for formal correctness
beforehand, so that we can use their tables during validation without
performing runtime checks to see if they are well-formed.

A complication arises when the MetaSchema validates the `default` field,
because it needs to use the definition of a field being currently
validated, which may in fact not be correct. Since the schema library
is written under the assumption that schema fields are indeed correct,
it can crash when validating `default` using a broken field definition.

Since the `"self"` code-path is used only in the MetaSchema validation of
default, it is reasonable to pay the cost of `pcall` here. Doing this
and allowing the rest of the library run under the assumption that the
schemas are always validated (which they currently are) allows us
to remove all runtime checks for formal validity of the schema tables
themselves, which is, in itself, a tiny performance win.