fix(core) fix check of declarative_config property
This is a complement to db28453c9.

When the `declarative_config` entry is set in `kong.conf` alongside
`database=postgres` or `database=cassandra`, this caused the initialization of
declarative config from DB-less mode to trigger incorrectly in DB mode,
causing problems with worker initialization. This commit fixes the issue.

Includes a regression test that demonstrates the problem.

Fixes #4508.