perf(*) plugins handlers do not have to inherit from BasePlugin anymore
Plugins handlers have always been inheriting from the `BasePlugin`
class, which is itself inheriting from the `Object` class provided by
`kong.vendor.classic`.

This design predates even the very first public release of Kong in 2015,
and was influenced by the OOP background of its original authors.

Today, the case can be made against this design:

- At best, it makes instances with `debug` log levels extremely noisy.
- At worst, it has a non-negligible performance impact due to the
  constant metatable lookups of a plugin's handlers methods and
  because of the non-JITable `ngx.log` calls sitting on hot code paths.
- All in all, its only purpose was to provide a `_name` attribute which
  was only used by the `BasePlugin` parent class and redundant with the
  `handler.name` property injected by the plugins loader.

This commit removes the need for plugins handlers to inherit from the
`BasePlugin` class, but still ensures backwards-compatibility for
plugins doing so.
Fixture plugins in our test suites were not updated and still inherit
from the `BasePlugin` class as a means to maintain regression tests for
this backwards-compatible behavior.

The `BasePlugin` is *not* deprecated by this commit.

In some initial benchmarkings (sample size 3), this change increased
req/s by up to 6%.