refactor(runloop) change the order of phase and ctx argument on plugins_iterator
Phase can never be nil so it makes sense to pass it as a second argument
and not as a third argument.

Context (ngx.ctx) is not used in init_worker with plugins iterator,
so it will be given as nil (in general we don't use ngx.ctx currently
on init_worker on Kong).

The ssl_certificate still uses ngx.ctx but there is issue with
OpenResty that the content of ngx.ctx is cleared on rewrite. This
change here removes mock, so it will start to work when upgrading
to OpenResty that has fixed this issue.

The next commit will fix the issue of init_worker not called on plugins
that implement it.