feat(pdk) make response.exit gRPC friendly
Changes kong.response.exit to not output body on gRPC requests. It will
instead output grpc-message header. It also output grpc-status on gRPC
responses. If plugin really wants to send a body it needs to specify
header Content-Type: application/grpc (or the Content-Type needs to be
prefixed with application/grpc). These changes rest on the fact that as
of now Kong doesn't support building and serving protobuf responses;
this will, however, change in the future and these assumptions will be
revisited and code updated accordingly.