fix(db_export) skip basicauth_credentials on db_export
This change makes basicauth_credentials non-exportable by db_export.

Explanation: basicauth_credentials have a built-in encryption system - the passwords are encrypted in the database.

If these are exported, they cannot be imported back to the database (they would be re-encrypted, rendering them useless).

The recommended way to import credentials for now is using plain, unencrypted passwords (which can not be obtained from the database).