feat(status) implement status_listen server
Problem
-------

Kong's Admin API exposes `/status` endpoint, which is consumed by other
services in the infrastructure to monitor Kong's health.

While this has worked for Kong for a long time,
we run into following issues with the current design:
- Users want to health-check Kong and also protect Kong's Admin API at
  the same time. If Kong's Admin API is protected using an auth plugin,
  health-checking becomes difficult. The reason being not all
  orchestration/platform tools performing health-checks support adding
  arbitrary headers to health-check requests.
- Users want to expose Kong's health outside a private network but keep
  the Admin API locked down.
- It is hard to perform HTTP based health-checks for a Kong node running
  only the proxy interface i.e. the Admin interface is disabled.
- A plugin could also potentially need to expose an endpoint which should not
  protected by any authentication.
  Prometheus plugin wants to expose `/metrics`, which should be not
  protected by any auth.
- Provision Kong Enterprise with RBAC enabled (health check won't pass):
  With Kong Enterprise, users want to enable RBAC from the get go,
  meaning when Kong is deployed, RBAC should be enabled. When this is
  done, health-checks for Kong using the current `/status` would fail
  (as we don't have any RBAC user while provisioning). This creates a
  chicken and egg where Kong needs to be up and running but that won't
  happen if status check is failing.

To work around this, users have resorted to injecting a custom server block
inside Kong using the NGINX directive injection feature, which adds
complexity and overhead in setting up Kong in a production environment.
Kong's Ingress Controller creates an additional file, mounts that as a volume
in the Docker container and then enables health-checking.

Health-checks and observability are common enough
(and simple for most deployments) that they should supported
out of the box and not require customizations.

Solution
--------

We introduce a `status` listener in Kong, which exposes insensitive health,
metrics and error read-only information from Kong.
Out of the box, only `/status` endpoint from Admin interface is available,
but plugins can add endpoints to this interface, as they can do for the Admin
interface.
This endpoint is strictly meant for machine-to-machine communication and
should not require any authentication/authorization.

The status server is disabled by default for
backwards-compatibility and not surprise existing users.