feat(router) support matching ports in Host header
Allow routes to specify a port in the host parameter. If present, it
matches only requests to that specific port. If absent, requests to
any port will match. Requests without explicit port are handled as if
they included the appropriate default port (80 for HTTP, 443 for
HTTPs). Routes with a port have higher priority than those without;
even for requests without an explicit port in the Host header.

From #5102