feat(router) add routes.path_handling attribute
This change introduces a new Routes attribute called path_handling.

It complements #5216 and #5227 by making it possible to select the
behavior the router will have when combining the Service Path, the Route
Path and the Request path into a single path sent to the upstream.

For now `path_handling` accepts two values:

* "v0" makes paths in the router behave like in Kong 0.x (like they did
  before #3780)
* "v1" makes paths behave like in Kong 1.x (like they did after #3780)

Notice that in order to add this new attribute a lot of tests needed to
be duplicated. The cases were v0 differs from v1 have been marked with
`-- !` for easy identification.