tests(rate-limiting) make tests less flaky with retrying
This performs the test up to two times (and no more than two).
We are **not** retrying to "give it another shot" in case of a flaky test.
The reason why we allow for a single retry in this test suite is because
tests are dependent on the value of the current minute. If the minute
flips during the test (i.e. going from 03:43:59 to 03:44:00), the result
will fail. Since each test takes less than a minute to run, running it
a second time right after that failure ensures that another flip will
not occur. If the second execution failed as well, this means that there
was an actual problem detected by the test.