refactor(conf) rename nginx_http_status_<directive> to nginx_status_<directive>
This commit renames injections for `http` `status` `server block` from
nginx_http_status_<directive> to nginx_status_<directive>.

This is done because we need to force injections to be in form of:
"nginx_[a-z]+_" OR `nginx_<single-word>_*`.

Otherwise we cannot reliably know where the context ends and where the
directive starts.

This commit does not add backward compatibility for old
`KONG_NGINX_HTTP_STATUS_<DIRECTIVE>`, because it overlaps with
`KONG_NGINX_HTTP_<DIRECTIVE>`. But this does add support for old
templates still looping around `nginx_http_status_directives`.