tests(helpers) add helper to load kong config without overrides
This allows, e.g., the Kong cached configuration (.kong_env file
located in the prefix) to be loaded without being overriden by
environment variables or defaults. This is useful, e.g., in the
`stop_kong` helper, which loads Kong configuration; allowing
overrides leads to issues when Kong was started (`kong_start`)
with configurations that differ from what is specified by env vars.

This issue was found in hybrid mode tests, which spin up a Kong
instance for the data plane, with `database=off`; when `stop_kong`
was called in CI and `KONG_TEST_DATABASE` was set to Postgres or
Cassandra, the call would fail (given dataplane mode can only be
used with database off).