feat(acl) 401 instead of 403 when no credentials found in context (#5440)
### Summary

This is an alternative approach to #5277 and #5236.

The background:

`acl` plugin runs before `request-termination` plugin, and `acl`
plugin only returned `403`. `403` seems like a correct status code
for authenticated consumers as they are not `unauthorized or 401`.
And also for anonymous consumers, if the consumer has `acl` groups
configured (which I have never seen outside our test suite). But
otherwise anonymous consumers should be getting `401`.

This PR tries to do almost the same as #5277 and #5236, but also
tries to not touch to tests too much and thus keeping the original
behavior intact (there is one test that needed to be changed).

All this work is from inspiration of Kong community, and I
(@bungle) was just a tool to polish the idea a bit.

Huge thanks to @jeremyjpj0916, and @odelvalle!

### Issues Resolved

Fix #4480