feat(db) squash 002_15_to_1 into 000_base
The main difference between 0.15 and 1.0 is the removal of the 'apis' entity,
but the migrations didn't actually remove the references to 'apis' from the
database itself. Here, since we are creating a base state for 1.0, we don't
need to create those unused references.

The actual `002_15_to_1` file concerned itself with removing `galileo`
plugins, which we also don't need to deal with in a base migration file.