feat(db) postgres to support different config for admin api and migrations
### Summary

Until recently we only had a single configuration for Postgres on our configs.
This changed when we decided to add `read-only` host support for it. This
expands that so that you can have different configuration for `admin api`
and for `migrations`. This allows Kong to use less privileged user on proxy
and more privileged user for `admin api` (needs to be able to write to
core-entities). Now migrations need even more powerful account that can also
do database schema changes.

This commit adds the following new configuration parameters:

- pg_admin_host = NONE
- pg_admin_port = NONE
- pg_admin_database = NONE
- pg_admin_schema = NONE
- pg_admin_timeout = NONE
- pg_admin_user = NONE
- pg_admin_password = NONE
- pg_admin_ssl = NONE
- pg_admin_ssl_verify = NONE
- pg_admin_max_concurrent_queries = NONE
- pg_admin_semaphore_timeout = NONE
- pg_migrations_host = NONE
- pg_migrations_port = NONE
- pg_migrations_database = NONE
- pg_migrations_schema = NONE
- pg_migrations_timeout = NONE
- pg_migrations_user = NONE
- pg_migrations_password = NONE
- pg_migrations_ssl = NONE
- pg_migrations_ssl_verify = NONE
- pg_migrations_max_concurrent_queries = 1
- pg_migrations_semaphore_timeout = NONE

If not specified, the value will be read from `pg_*` (the place where
you usually configure the proxy configuration for write access).