feat(config) add "kic" configuration for reporting (#5765)
* feat(config) add "kic" setting

Add a "kic" setting to kong.conf. This is a boolean indicating whether
the instance is managed by the Kong Ingress Controller. It is not
documented and not included in kong.conf.default, as users should not
set it directly. Kubernetes tooling will set it automatically if
appropriate.

* feat(reports) send KIC usage

Add ingress controller usage setting to anonymous reports.