tests(balancer) add a test case covering removal/addition of an identical target
This test case covers the issue fixed by
5d8c87959cad18ae82518003199df010bc2c6a5c and recently investigated in
FTI-1458.

In the investigated issue, this invalid balancer state was reproduced in
a node receiving updates from the cluster events. In this test case, we
reproduce this state directly by querying the proxy port of the same
node (i.e. the node receiving Admin API requests).