fix(declarative) when loading declarative config, remove null values properly
When declarative config with `null` value is passed in YAML or JSON, the
lyaml parser will return it's own variant of `null` representation that
is different from cJSON. Causing schema validations to fail with type
error.

This commit adds `lyaml.null` support to the `remove_nulls` helper to
make sure they are always properly removed from the input.

This commit also removes the use of `cjson` when parsing JSON input and
instead always use the lyaml library. Because when user specifies a
string payload inside the `config` parameter when POSTing to the
`/config` admin API endpoint, the lyaml parser has always been used.
Note the [YAML Specification](https://yaml.org/spec/1.2/spec.html)
specifically calls "The primary objective of this revision is to bring
YAML into compliance with JSON as an official subset."

Because JSON is a subset of YAML and libyaml supports parsing JSON out
of the box, we can use lyaml for both YAML and JSON config when parsing
declarative config. This avoids different code path when loading JSON
encoded declarative config using the CLI and using the admin API to make
sure they always present the same result and fail with the same messages
for the user.

fixes #5834